Added integrators trait
